+++ src/Server/ServerNetwork/Network/Infrastructure/ServerNetworkManager.cs
@@ -131,18 +131,24 @@ public sealed class ServerNetworkManager : INetworkManager
 
     public void Poll()
     {
-        // Handle incoming messages, connections, etc.
-        netManager.PollEvents();
-
         // Send whatever is in the outbound event queue.
         while (outboundEventQueue.TryDequeue(out var evInfo))
         {
+            // TODO Send event
         }
+
+        // Handle incoming messages, connections, etc.
+        netManager.PollEvents();
     }
 
     public void EnqueueEvent(OutboundEventInfo evInfo)
     {
         outboundEventQueue.Enqueue(evInfo);
+    }
+
+    public void Disconnect(int connectionId)
+    {
+        // TODO
     }
 
     /// <summary>