+++ src/Server/ServerNetwork/Network/ServerNetwork/ServerNetworkSystem.cs
@@ -1,0 +1,88 @@
+// Sovereign Engine
+// Copyright (c) 2023 opticfluorine
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a
+// copy of this software and associated documentation files (the "Software"),
+// to deal in the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// and/or sell copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS IN THE SOFTWARE.
+
+using System.Collections.Generic;
+using Sovereign.EngineCore.Events;
+using Sovereign.EngineCore.Systems;
+using Sovereign.NetworkCore.Network.Infrastructure;
+using Sovereign.ServerCore.Events;
+
+namespace Sovereign.ServerNetwork.Network.ServerNetwork;
+
+/// <summary>
+///     Responsible for managing network resources at the server.
+/// </summary>
+public class ServerNetworkSystem : ISystem
+{
+    private readonly INetworkManager networkManager;
+
+    public ServerNetworkSystem(EventCommunicator eventCommunicator, IEventLoop eventLoop,
+        INetworkManager networkManager)
+    {
+        this.networkManager = networkManager;
+        EventCommunicator = eventCommunicator;
+        eventLoop.RegisterSystem(this);
+    }
+
+    public EventCommunicator EventCommunicator { get; }
+
+    public ISet<EventId> EventIdsOfInterest => new HashSet<EventId>
+    {
+        EventId.Server_Network_DisconnectClient
+    };
+
+    public int WorkloadEstimate => 5;
+
+    public void Initialize()
+    {
+    }
+
+    public void Cleanup()
+    {
+    }
+
+    public int ExecuteOnce()
+    {
+        var eventsProcessed = 0;
+        while (EventCommunicator.GetIncomingEvent(out var ev))
+        {
+            switch (ev.EventId)
+            {
+                case EventId.Server_Network_DisconnectClient:
+                    OnDisconnectClient((ConnectionIdEventDetails)ev.EventDetails);
+                    break;
+            }
+
+            eventsProcessed++;
+        }
+
+        return eventsProcessed;
+    }
+
+    /// <summary>
+    ///     Handles a request to disconnect the client associated with the given connection.
+    /// </summary>
+    /// <param name="details">Request.</param>
+    private void OnDisconnectClient(ConnectionIdEventDetails details)
+    {
+        networkManager.Disconnect(details.ConnectionId);
+    }
+}